Refaktorerte koden
